Refatorado modulo 'views'
